changed handlers

+++ Examples/MAX78000/Stepper_Control/flags.c
@@ -1,10 +1,10 @@
 #include "flags.h"
 
 // stores the flag bits
-volatile int flags = 0;
+volatile uint32_t flags = 0;
 
-void set_flag(Flag f) { flags |= 1 << f; }
+void set_flag(Flag f) { flags |= (uint32_t)1 << f; }
 
-void unset_flag(Flag f) { flags &= ~1 << f; }
+void unset_flag(Flag f) { flags &= ~(uint32_t)1 << f; }
 
-bool is_flag_set(Flag f) { return flags & (1 << f); }
+bool is_flag_set(Flag f) { return flags & ((uint32_t)1 << f); }
